Room index api and show api point +tests

+++ database/migrations/2023_04_16_094336_create_rooms_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
         Schema::create('rooms', function (Blueprint $table) {
             $table->id();
             $table->foreignId('branch_id');
-            $table->tinyInteger('room_number');
+            $table->integer('room_number');
             $table->tinyInteger('room_floor_number');
             $table->tinyInteger('room_status')->default(1); //may be room is not approved by admin for reservation due to any faults
             $table->boolean('hidden')->default(false);
